MultipleParallelImportMetadataSourceServiceImpl.java
This is what the University of Bamberg
uses (https://fis.uni-bamberg.de) to look up DOIs specified by the
user for import in both the Crossref registry and the DataCite
registry, in parallel.

This source can be specified just once, replacing the Crossref and
DataCite sources by a single entry.