Adjustments to POM files so that the changes in "modules/server" are incorporated into the JAR generated by "server-boot".
This commit adds an "attachClasses" parameter to the "maven-war-plugin"
in the "dspace/modules/server/pom.xml", which generates a JAR file that
can be referenced in the "dspace/modules/server-boot/pom.xml"
(see <https://maven.apache.org/plugins/maven-war-plugin/war-mojo.html#attachClasses>)
via

```
        <dependency>
            <groupId>org.dspace.modules</groupId>
            <artifactId>server</artifactId>
            <classifier>classes</classifier>
        </dependency>
```

The dependency must be placed *before* the "dspace-server-webapp"
dependency, to ensure that it overrides the classes in the
"dspace-server-webapp" module. In the "server-boot.jar", the
CLASSPATH is determined by the order of JARs in the
"BOOT-INF/classpath.idx", which is generated based on the order of
dependencies in the POM (see https://stackoverflow.com/a/67997782).

The root "pom.xml" file was modified to provide the version for
"modules/server" JAR file, in keeping with how the versions of other
JAR files are specified.

(cherry picked from commit b45e7cdac5147d3936c020ffc20f7dd30e817fea)